few fixes

+++ Assets/App/Scripts/Cpu/Clocks.cs
@@ -6,12 +6,12 @@ namespace StudioKurage.Emulator.Gameboy
 {
     public partial class Cpu
     {
-        public int mc;  // machin cycle
-        public int cc;  // clock cycle
-        public int imc; // machin cycle of last instruction
-        public int icc; // clock cycle of last instruction
+        public long mc;  // machin cycle
+        public long cc;  // clock cycle
+        public long imc; // machin cycle of last instruction
+        public long icc; // clock cycle of last instruction
 
-        public int m {
+        public long m {
             set {
                 imc = value;
                 icc = value * 4; // always equals m * 4
